requirements.txt
google-cloud-aiplatform[adk,agent_engines]
google-cloud-geminidataanalytics
google-auth
google-adk
flask
flask-cors
